Something growing v0.1
Most of the work was done on the p5 sandbox website, hence the big commit. Summarizing:

It creates a tree, yes an actual tree by creating a branch (a line basically)
and then growing additional branches from it based on parameters such as max depth and no of branches
it will also stop if max no branches is reached. Thickness indicates thickness of lines but also
opacity which becomes less and less the deeper you go in the branches.